Change workshop item scaling type to a real number

+++ workshop_info.h
@@ -9,6 +9,6 @@ struct WorkshopInfo {
   string SERIAL(verb) = "produces";
   AttrType SERIAL(attr);
   int SERIAL(minAttr) = 1;
-  int SERIAL(itemScaling) = 1;
+  double SERIAL(itemScaling) = 1;
   SERIALIZE_ALL(NAMED(furniture), NAMED(name), OPTION(verb), NAMED(attr), OPTION(minAttr), OPTION(itemScaling))
 };
